added shortcode functionality
After not making much progress with custom fields I looked into the shortcode side of the task to see if I could make much progress there.

Found some helpful discussion threads which made the process much simpler. it may not be perfect but it's close to how it came to my mind when reading the tech challenge.